cli: Refactor `graph-build`
Split `args` module which can be reused
for `graph-deploy`.
